cleanup of toilet routes - change of structure
