+++ src/api/v1/controllers/toilet/feature.controller.ts
@@ -1,8 +1,8 @@
 import { Request, Response, NextFunction } from "express";
-import db from "../../../../config/database";
-import { FeatureSchema } from "../../schemas/toilet.schema";
+import db from "@config/database";
+import { ToiletFeature } from "../../schemas/toilet/toilet.entity.schema";
 
 export const getAllFeatures = async (req: Request, res: Response) => {
-  const allFeatures: FeatureSchema[] = await db.query.feature.findMany();
+  const allFeatures: ToiletFeature[] = await db.query.feature.findMany();
   res.status(200).json(allFeatures);
 };
