+++ src/api/v1/schemas/auth/auth.request.schema.ts
@@ -2,6 +2,7 @@ import { z } from "zod";
 import { extendZodWithOpenApi } from "@asteasolutions/zod-to-openapi";
 extendZodWithOpenApi(z);
 
+//REGISTER
 export const registerRequest = z
   .object({
     email: z
@@ -21,17 +22,7 @@ export const registerRequest = z
   .openapi({ description: "Schema for register request" });
 export type RegisterRequest = z.infer<typeof registerRequest>;
 
-export const registerResponse = z.object({
-  message: z.string().openapi({
-    description: "Message about the server response",
-    example: "Registration Succesfull",
-  }),
-  user: z.object({
-    email: z.string().email().openapi({ example: "example@gmail.com" }),
-  }),
-});
-export type RegisterResponse = z.infer<typeof registerResponse>;
-
+//LOGIN
 export const loginRequest = z.object({
   email: z
     .string()
@@ -47,27 +38,8 @@ export const loginRequest = z.object({
 });
 export type LoginRequest = z.infer<typeof loginRequest>;
 
-export const loginResponse = z.object({
-  message: z.string().openapi({ example: "Login succesfull" }),
-  token: z.string().openapi({ example: "your_jwt_token_string" }),
-  refreshToken: z
-    .string()
-    .openapi({ example: "your_jwt_refresh_token_string" }),
-  user: z.object({
-    id: z.number().openapi({ example: 123 }),
-    email: z.string().openapi({ example: "example@gmail.com" }),
-  }),
-});
-export type LoginResponse = z.infer<typeof loginResponse>;
-
+//REFRESH TOKEN
 export const refreshTokenRequest = z.object({
   refreshToken: z.string().openapi({ example: "your_jwt_token_string" }),
 });
 export type RefreshTokenRequest = z.infer<typeof refreshTokenRequest>;
-
-export const refreshTokenResponse = z.object({
-  message: z.string().openapi({ example: "Refreshing token sucessfull" }),
-  token: z.string().openapi({ example: "your new jwt access token" }),
-  refreshToken: z.string().openapi({ example: "your new jwt refresh token" }),
-});
-export type RefreshTokenResponse = z.infer<typeof refreshTokenResponse>;
