+++ src/api/v1/types/types.ts
@@ -1,0 +1,6 @@
+import { getToiletById } from "../repos/toilet.repo";
+
+export type ToiletWithRelations = Awaited<ReturnType<typeof getToiletById>>;
+export type ToiletWithRelationsAndDistance = ToiletWithRelations & {
+  distance: number | null;
+};
