+++ tsconfig.json
@@ -17,8 +17,19 @@
       "jsonwebtoken",
       "passport",
       "passport-google-oauth20"
-    ]
+    ],
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["src/*"],
+      "@config/*":["src/config/*"], 
+      "@db/*":["src/db/*"],
+      "@middleware/*": ["src/middleware/*"],
+      "@services/*": ["src/services/*"],
+      "@v1/*": ["src/api/v1/*"],
+      "@error/*":["src/errors/*"]
+    }
+
   },
-  "include": ["src/**/*", "src/types/express.d.ts"],
+  "include": ["src/**/**/*", "src/types/express.d.ts"],
   "exclude": ["node_modules", "dist"]
 }
